[refactor]: service-worker 함수 별도파일 분리
