+++ c-v/public/manifest.json
@@ -10,7 +10,8 @@
     "128": "icon/icon-128.png"
   },
   "background": {
-    "service_worker": "service-worker.js"
+    "service_worker": "service-worker.js",
+    "type": "module"
   },
   "action": {
     "default_popup": "index.html",
