+++ c-v/public/modules/getListStore.js
@@ -1,0 +1,11 @@
+import openDatabase from './openDatabase.js';
+
+const getListStore = async (permission) => {
+  const db = await openDatabase();
+  const transaction = db.transaction(['list'], permission);
+  const listStore = transaction.objectStore('list');
+
+  return listStore;
+};
+
+export default getListStore;
