+++ c-v/public/modules/openDatabase.js
@@ -1,0 +1,19 @@
+const openDatabase = () => {
+  return new Promise((resolve, reject) => {
+    const request = indexedDB.open('CVStore', 1);
+
+    request.onupgradeneeded = (event) => {
+      const db = event.target.result;
+
+      if (!db.objectStoreNames.contains('list')) {
+        const listStore = db.createObjectStore('list', { autoIncrement: true });
+        listStore.createIndex('name', 'name', { unique: false });
+      }
+    };
+
+    request.onsuccess = () => resolve(request.result);
+    request.onerror = () => reject('Failed to open database');
+  });
+};
+
+export default openDatabase;
