+++ c-v/public/modules/service/editCommand.js
@@ -1,0 +1,26 @@
+import getListStore from '../getListStore.js';
+import getListByName from '../getListByName.js';
+import getPrimaryKey from '../getPrimaryKey.js';
+
+const editCommand = async (currentListName, targetCommand, newCommand) => {
+  try {
+    const listStore = await getListStore('readwrite');
+    const nameIndex = listStore.index('name');
+
+    const currentList = await getListByName(currentListName, nameIndex);
+
+    const primaryKey = await getPrimaryKey(currentListName, nameIndex);
+
+    const targetIndex = currentList.commands.findIndex(
+      (commandItem) => commandItem === targetCommand,
+    );
+
+    currentList.commands[targetIndex] = newCommand;
+
+    await listStore.put(currentList, primaryKey);
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+export default editCommand;
